+++ src/Dialog/DialogTitle.js
@@ -15,7 +15,7 @@ DialogTitle.propTypes = {
     className: PropTypes.string,
     component: PropTypes.oneOfType([
         PropTypes.string,
-        PropTypes.element,
+        PropTypes.elementType,
         PropTypes.func
     ])
 };