+++ src/Grid/Grid.js
@@ -8,7 +8,7 @@ const propTypes = {
     className: PropTypes.string,
     component: PropTypes.oneOfType([
         PropTypes.string,
-        PropTypes.element,
+        PropTypes.elementType,
         PropTypes.func
     ]),
     noSpacing: PropTypes.bool,