+++ src/Layout/Content.js
@@ -17,7 +17,7 @@ Content.propTypes = {
     className: PropTypes.string,
     component: PropTypes.oneOfType([
         PropTypes.string,
-        PropTypes.element,
+        PropTypes.elementType,
         PropTypes.func
     ])
 };