+++ tests.setup.js
@@ -1,0 +1,3 @@
+const Enzyme = require('enzyme');
+const Adapter = require('enzyme-adapter-react-16');
+Enzyme.configure({ adapter: new Adapter() });