+++ tests.webpack.js
@@ -1,3 +1,5 @@
+require('./tests.setup');
+
 // see https://github.com/webpack/karma-webpack#alternative-usage
 
 const context = require.context('./src', true, /-test\.js$/);