+++ webpack.config.js
@@ -13,7 +13,7 @@ module.exports = {
             amd: 'react-dom'
         }
     },
-    devtool: 'sourcemap',
+    devtool: 'source-map',
     module: {
         loaders: [
             { test: /\.js$/, loader: 'babel-loader', exclude: /(node_modules)/ }